Stdlib benchmarks and tests
Removed the tests and benchmarks from main.go and used the go stdlib testing package for tests and benchmarks.

To run tests and benchmarks use:
    test -v
or
    test -v -bench=.
